alex/style-followups-2: fill defs come from shapes
